requirements.txt
affine==2.4.0
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
imageio==2.31.2
japanize-matplotlib==1.1.3
joblib==1.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
opencv-python==4.8.0.76
pandas==1.3.5
patsy==0.5.4
pillow==9.5.0
pykwalify==1.8.0
pyparsing==3.1.1
pyradiomics
pytz==2023.3.post1
pywavelets==1.3.0
pyyaml==6.0.1
ruamel-yaml==0.17.32
ruamel-yaml-clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-posthocs==0.8.0
scipy==1.7.3
seaborn==0.12.2
simpleitk==2.2.1
statsmodels==0.13.5
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.66.1
typing-extensions==4.7.1
xgboost==1.6.2